job-application-system/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.7.4
python-dotenv==1.0.1
bcrypt==4.2.0
pydantic[email]==2.8.2
PyJWT==2.8.0
fastapi-mail==1.4.1
python-multipart==0.0.9
redis==5.2.1
pytest==8.3.3
httpx==0.27.2
requests==2.32.3
websockets==12.0
